Initial commit of refactor to consolidate to single GenericListView component
Major refactor. Remove all the duplicated code and introduce some separation of concerns.

GenericListView is what is says it is. You should just need to specify the field, headerName, and a cell rendering function in the model and all the lists should behave fine.

The cellrenderers will need more consolidation.

Need to reintroduce various additional features.